requirements.txt
streamlit
tensorflow_hub
numpy
tensorflow-cpu